Day 3!
Only had a little time to complete the homework, but since the video gave a really good insight and explanation to things like "states" and "effects" it was super easy to create the things needed that used the firebase database and the components with files linking to each other. This is honestly really fun working wit react and I am learning a lot.